requirements.txt
google-genai
pyyaml
requests
python-dotenv
pytest
pytest-mock
